Re-word "Disallow USB File Transfer" manual test.
The existing manual test instructions assume that the device is capable of
USB file transfer using MTP or PTP, and aim to ensure that disallowing
file transfer indeed blocks that option.

The new instructions allow for devices that don't support MTP or PTP at
all, and also clarify that 'adb pull' doesn't count as "USB file
transfer" in this context.

Bug: 32335136
Change-Id: Iab4eaa13b116871876ad5a590136f54e58029fc8

Former-commit-id: 87aba921bbe682a2dd842e43a02bdac5fc2df10b