Add NativeClearKeySystemTest test. *DO NOT MERGE*
Merge from go/ag/1598039

For os version below 8, this test needs go/ag/1628977 and
go/ag/1598174 which may not be available for devices that
are in the fields. isVersionSmaller() checks and skip this
test if version is below 8. This test is no longer needed
for O and up, so the function is removed.

Tests MediaDrm NDK APIs. ClearKey system implements a subset of NDK
APIs, this test only tests the APIs that are supported by ClearKey
system.

Test: ANDROID_BUILD_TOP= ./android-cts/tools/cts-tradefed run cts -m CtsMediaTestCases --test android.media.cts.NativeClearKeySystemTest#testClearKeyPlaybackCenc

Test: ANDROID_BUILD_TOP= ./android-cts/tools/cts-tradefed run cts -m CtsMediaTestCases --test android.media.cts.NativeClearKeySystemTest

Test: Test passes for API level 25 devices.

bug: 17255611
Change-Id: I8998963ca35723af74a1263c337db893923436a8

Former-commit-id: 80428c9997423754ec5d996809e988125df300ab