Check that H.264 is available before checking max frame rate
Previously this test was checking for the max frame rate of H.264
before it would check if the H.264 codec actually existed on the
device. Devices that do not have this codec would fail the test
due to this ordering. This change reorders the checks so that
we only check the max frame rate if the code exists.

Bug: 32742421
Change-Id: I3a2e0806c1d6dd87095eadd7917d25273acfe190

Former-commit-id: a175c5e8ced5a4796215ee66ffb8678c99dc401a