Fix testGetWindowVisibleDisplayFrame in multiwindow mode
When mAttachInfo is null, View.getWindowVisibleDisplayFrame
returns the size of the real display (adjusted for system
decorations). This causes the test to fail in multiwindow
mode when the activity-specific display is smaller than the
real display.

Fix this by using the application context to retrieve the
real size of the display instead.

Bug: 28850117
Test: run cts-dev -d --module CtsViewTestCases --test
android.view.cts.ViewTest#testGetWindowVisibleDisplayFrame
passes in both fullscreen and multi-window mode.

Change-Id: Ibffd65c0ca1d8bd8637202add5451f40ea73df40

Former-commit-id: 44c245f5d17bedf8e4e57d78bb83aac11d784157