DO NOT MERGE
Remove ClonedSecureRandomTest

This test was kind of flaky because it depended on a lot of variables
working. SELinux has now blocked pid_max which makes this test fail 100%
of the time.

It is obsoleted by BoringSSL. This problem only affected OpenSSL because
it buffered random data which could be a problem when forking.

Test: cts-tradefed run cts -m CtsSecurityTestCases -a arm64-v8a
Bug: 34077876
Change-Id: I9ea90a8f1e1620d184fd0315ea119abb601c9658

Former-commit-id: 3ad6d3b48db3011d0b6b35e12fad31371cb8d452