Create @VendorInterfaceTest annotation and provide examples
The annotation can be applied at the testcase and the test class
level. The annotation is not sufficient for signalling that the
test should be run as a part of cts-vendor-interface, and an
include filter for the module (e.g. CtsAccelerationTestCases) is
also necessary.

bug: 36423839
Test: run cts-vendor-interface
merged-in: I63c11c6f7f3a57ec2f1264b1c34d7a0e8bb93363
Change-Id: Ic3115af5ae2c265ad869304f319654fe9cca3694

Former-commit-id: f4298ddb24b273ef8fc52e14313144347f3af87b