Fix the sync adapter account access CTS tests
Accounts are accessed by sync adapters singed with the same
cert as the authenticator managing the account, otherwise the
user has to approve account access. The CTS tests were failing
because the sync manager updates its internal state on an
accounts change via a broadcast which takes time to propagate.
The fix is to wait a bit for the propagation to occur which
unfortunately is not detectable, so a hard-coded wait is used.
Also these tests would not pass if there is no network
connection as no syncs happen if there is no connectivity.

Test: These tests are now passing

bug:36215841

Change-Id: I5d8374802dc10bbd3a597175dce4f4942234a4c4

Former-commit-id: 18c94d6f87885cbc3973feb3b9b2b23f51de4a76