monsoon: Add option to collect USB power data
This adds the ability to collect power consumption data from the USB
passthrough channel in addition to the main power channel.

To avoid breaking compatibility with tools that parse monsoon.py's
output, a new flag, --includeusb, is added. When it isn't specified,
the output is unchanged.

It is not implemented here, but I've tried to code this in such a way
as to make it easy to add collection from the auxiliary channel too,
if such a need should arise.

Change-Id: Ia1ce3d5350fdbacfccdc404b04fcab734d56f544

Former-commit-id: c3520a96b7944c4c3b926fdb397c0a95db3886d5