Hostside smarter about waiting for test to finish
For Batterystats CTS tests that use the BgVsFgActions method of
performing actions: previously, the hostside would just sleep while
waiting for the action to be performed on the device, guessing how long
it might take and sleeping for that length of time. Now the device will
send a message via the logcat to indicate that the action is finished,
so the hostside part of the test can scan the logcat to determine when
the action is done (instead of sleeping an arbitrary length of time).

Test: run cts-dev -m CtsIncidentHostTestCases -t
com.android.server.cts.BatteryStatsValidationTest

Change-Id: I0d646aee2b4b0e6f5a6ebefca4e5b55062a05f58

Former-commit-id: a6685d3f6a7833d862c0f20ced35d136e63b432f