Download and Upload tests stay in bg or fg
For the Batterystats CTS tests, previously the testWifiDownload and
testWifiUpload ran their task async, which means that a foreground
activity could finish while the download/upload was occurring. For the
test, it would be better for the proc stay to stay constant for the
duration of the test.

Test: run cts-dev -m CtsIncidentHostTestCases -t
com.android.server.cts.BatteryStatsValidationTest

Change-Id: I00bfa96c8e123479b57bd2887a65ce3050ea2a81

Former-commit-id: 42b961f780ef6eb53ad0264f67033478417499ba