Additional CTS Test for noHistory flag.
This test ensures that the noHistory flag is honored, even if the
resuming activity does not enter idle. To exercise this constraint,
we start an activity known to not enter idle on top of the noHistory
activity and press back. The expected behavior is that the noHistory
activity will be gone and the launcher will be shown.

Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test CtsServicesHostTestCases android.server.cts.ActivityManagerActivityVisibilityTests#testNoHistoryActivityFinishedResumedActivityNotIdle
Bug: 37234980

Change-Id: I4ccde8428ec0c6bffb2b0614eeb48bc4e0fa6e9c

Former-commit-id: 68f5a2a5d36542d443ddc62c213914d877ca8ec8