Reduce flakiness of a11y testTextEditingActions
Putting back a few lines of this test that were removed in
ag/270663. I'm guessing they were removed as unnecessary
because the copy action cleared the selection, but that
action's behavior seems not to be guaranteed (ag/876521).
Insulating us from the behavior of the widgets seems like
a good idea, and it also seems to fix the flake.

Fixes: 38113232
Test: Ran this test. It failed 3 times in a row without this
change, then passed 5 times in a row with it.

Change-Id: I913664767c65dc23ee295fd8dcb5f13b67a17efc

Former-commit-id: faba6cf04096a2a02336f08b459373377835ad59