Verify incoming call handled by specified default dialer.
Test: manual
Bug: 35338645

Outgoing calls already tested by BasicInCallServiceTest.java
This test implicitly tests that the user can change the default dialer.

The test is as follows:
1) The user is prompted to change the default dialer.
2) The user is prompted to call the device.
3) We provide an InCallService as the new default dialer; this is to
avoid any confusion between manually specified dialer variants.
4) The telecom stack binds to our default service when it receives the
incoming call.
5) We register that we received the incoming call in our service, and
update the relevant state in the test.
6) The user can now check the "activity pass" button.
7) The user can reset the default dialer.

The intent of the test is to veryify that a third party dialer is
capable of recieving incoming calls from the Telephony system by acting
as an incoming call InCallService UI.

Change-Id: Ibeca66c132f14d136a018740254fbe684c94b68b

Former-commit-id: ec56cb3bbd89029723ad35bab99ebfa719b1dad0