CTS test for 'fullBackupOnly' manifest attribute
Uses 3 different versions of the same app that differ only by
'fullBackupOnly' value and the presence/absence of the backup agent.

Tests the following scenarios:

1. App that doesn't have fullBackupOnly (same as fullBackupOnly=false by
default) and has a backup agent will get key/value backup.
We check that key/value data was restored after reinstall and dolly data was not.

2. App that doesn't have fullBackupOnly (same as fullBackupOnly=false by
default) and has no backup agent will get Dolly backup.
We check that key/value data was not restored after reinstall and dolly data was.

3. App that has fullBackupOnly=true  and has a backup agent will only
get Dolly backup.
We check that key/value data was not restored after reinstall and dolly data was.

The flow of the tests is the following:
1. Install the app
2. Generate files in data folder of the app, including a file in
no_backup folder. The file in no_backup folder is used for key/value backup by
the backup agent of the app.
3. Run 'bmgr backupnow'. Depending on the manifest of the app we expect
either Dolly or Key/value backup to be performed.
4. Uninstall and reinstall the app.
5. Check that the correct files were restored depending on the manifest:
 - Files in the app's data folder for Dolly backup
 - Only file in no_backup folder for key/value.

Test: cts-tradefed run cts-dev -m CtsBackupHostTestCases -t android.cts.backup.FullBackupOnlyHostSideTest

Bug: 38108587

Change-Id: If29577a5209fc77cebebe23dc879c51ab4be5e65

Former-commit-id: c131913926919e5237b7f6d10f96239277b73257