Partially fix CTS Preference tests for wearable devices.
This CL tackles one of the many problems there are with this test on
wearable devices. In particular it ensures that while searching for
a particular view the container view is scrolled before it is concluded
that the view does not exist. In this test, the size of the screen was
not taken into account so on small screens of wearable devices it would
not find elements of the list that are not yet loaded (as they are off
screen).

Bug:37646009
Test: Manually ran tests on Angelfish and Pixel devices

Change-Id: Ia39a97899ede9c5820c20e251ccaa9b4910ad6df

Former-commit-id: 28cdb05921e02663e117dc46187cfa7f60c41d94