Increates timeouts and add a retry rule
Unfortunately we don't have a way to know when the account
chagne broadcast is delivered to the sync manager so need
to have a timeout. This change increases this timeout and
also adds a retry rule to make this test more robust. The
test passes locally 100% so hopefully this fixes the prod
failures.

Test: SyncAdapterAccountAccessHostTest pass

bug:62390229

Change-Id: I6cffa9f329a79968b69b4b298175a0269e0b1532

Former-commit-id: 6ab5b5fef9841d48c4ef69538447da308a779e3c