Fix flakiness of testReportShortcutUsed due to async event population.
It seems that the previous attempt I6624a9af4030c70ba527fe2a5416e3e4204d3dee
to fix the flakiness of the case introduced another class
of flakiness. The current change is to fix it.

Due to the asynchronous nature, each queryEvents() paired
with reportShortcutUsed() must wait for the event to be
populated, not only for the first call for |id3| but
also for the rest (|id1|, |idManifest|).

Bug: 62473595
Test: CtsShortcutManagerTestCases
Change-Id: I940418e465e3bc6ec0366b9322d4edad0991561f

Former-commit-id: 9bcb7992e965a167a8c1a6125052fe81c660c2cf