Add Verifier test for always-on VPN Settings UI.
We add a CTS Verifier test to make sure that the Settings UI is
implemented correctly for the always-on VPN feature. This test verifies
the following:

1) VPN apps targeting SDK 23 can't be set as always-on in Settings (to
catch bugs like b/35416283)

2) VPN apps targeting SDK 24 can be set as always-on in Settings

3) VPN apps that opt out of the always-on feature can't be set as
always-on in Settings

This test is located under CTS Verifier > Managed Provisioning
> BYOD Managed Provisioning > Always-on VPN Settings.

Although the feature being tested works in both the primary user and the
managed profile, we only test it in the managed profile, with the
assumption that if it's implemented correctly in the managed profile, it
should also work in the primary user.

Bug: 35668087
Test: make cts-verifier -j and follow instructions
Change-Id: I84e714960f8e84918daf37feff022a9e60bbde38

Former-commit-id: 89bda13c394badf93d0eb3d852b521527a451670