Fixes getting the wrong app transition going home after minimized (2/2)
This is a side affect from ag/2125707. When minimizing and unminimizing
back to home would cause the same affect in the bug, where the app
transition is from activity to activity and not app. The issue was a
"updateAllDrawn" is normally called at the end of the surface
transaction and therefore removing this call (that would run inbetween
the transaction) should fix prematurely calculating the app transition
before setting the wallpaper target.

Fixes: 62727547
Test: go/wm-smoke
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test
CtsServicesHostTestCases
android.server.cts.ActivityManagerDockedStackTests#
testMinimizeAndUnminimizeThenGoingHome
Change-Id: I5a0f2d6087640ad17fa2cf1a19840e17aa166151

Former-commit-id: 8fcd470faad99fe9808d77e6574b2562d6740aad