Progress towards FBE and adoptable storage.
Adoptable storage is a feature present on many Android devices with
physical SD card slots or stable USB OTG ports.  We want to give
developers a way to validate and debug their apps on all device
styles, regardless of what that device physically supports, so we
added a new "virtual disk" feature that presents itself as a
physical SD card for testing purposes.

This is also valuable for devices that only have a single USB OTG
port, since it required an awkward CTS testing setup with adb-over-
network.  Using this virtual disk feature we can validate the
behavior over a normal adb connection.

Also extend the test to run over mulitple users, when supported.

Since we can race ahead of PackageManager scanning and loading of
newly mounted volumes, use a new dump feature to explicitly watch
for the volume to finish being loaded/unloaded.

Test: cts-tradefed run commandAndExit cts-dev --abi armeabi-v7a -m CtsAppSecurityHostTestCases -t android.appsecurity.cts.AdoptableHostTest
Bug: 37436961, 29923055, 25861755, 30230655
Change-Id: Ie5be88429b7e182abeffed80b0d60612dd72c3a9

Former-commit-id: bc001dab575c60fb233b47abbe659d91c77a9b21