Add test preconditions to hostside backup CTS.
Check that backup enabled and local transport is selected before each
test. We switch backup on and select LocalTransport in BackupPreparer
before all the module is run, so if at any time the state is different
from what we expect, that means something went wrong and we want to
fail.

Test: make cts -j40 && cts-tradefed run cts -m CtsBackupHostTestCases
Bug: 62253989
Change-Id: I5f47e44d44b05cb5aef7ec9e36130d2391f41fe8

Former-commit-id: b4d7a2ccf0e42335a5f8139b7840808340cd5525