Make AlmostFullTest no longer overfill the disk.
createNewFilledFile currently creates a file that is rounded up to
the nearest FileUtil.BUFFER_SIZE bytes, and AlmostFullTest creates
a series of files of files that are not an integral multiple of that
value. Hence, for a large number of filler files, the test overfills
the storage device and fails.

This change works around the disparity by having AlmostFullTest use
an integral multiple of FileUtil.BUFFER_SIZE bytes for the filler
files, without affecting the other tests.

Bug: 63535343
Test: run cts --module CtsFileSystemTestCases
Change-Id: I44e6d7869f2f1def334efca21153e22dcfa37167

Former-commit-id: 8e8a4bd7173e8b69067d5bfc7b1d267ecdda36b5