sphal is another linkernamespace that is available to JNI libs
CTS tries to dlopen every libs under /system/lib and checks if the
result of dlopen is as expected. If a lib name is not mentioned in
/system/etc/public.libraries.txt, then the dlopen should fail with a
message like 'library xxxx is not accessible for the namespace
"classloader-namespace"'. CTS compares the result of dlerror() with this
string literal.

However, there is a corner case that produces different error message;
libX.so is both in /system/lib and /vendor/lib, and libX.so is added to
/vendor/etc/public.libraries.txt. In this case, linker tries to load
/system/lib/libX.so in the 'sphal' namespace (since it uses file's
base name as its soname when crosssing the namespace boundary). As a
result the error message is like '... for the namespace "sphal"'.

So, adding '... for the namsapce "sphal"' as one of the expected error
message for the case.

Bug: 63974074
Test: put libX.so to both /system/lib and /vendor/lib.
add libX.so to /vendor/etc/public.libraries.txt.
android.jni.cts.JniStaticTest#test_linker_namespaces passes.

(cherry picked from commit 91db7d33a2e8b34bdd2133cb3c59ae7ba709f69d [formerly c7a3e59b56e29e9331f0028f4cb730eb5ea6acda])

Change-Id: Ib5cdf46ed917833036dde449c7b36d8dbd7f0b53

Former-commit-id: a20ad3b000d6e7237fd8f5644c0482f2d43e4bee