Tests for EditText last line spacing
StaticLayout did not work as expected when there is a line ending with
'\n'. For those cases it would calculate 2 lines, but would
discard line spacing for both of them. This CL adds a test for the fix
of that case, and also updates the DynamicLayout test to expect the
same values for line spacing as the StaticLayout. Previously the last
line was expected to have line spacing.

Test: bit -t CtsTextTestCases:android.text.cts.StaticLayoutTest
Test: bit -t CtsTextTestCases:android.text.cts.DynamicLayoutTest

Bug: 25194907
Change-Id: I68fa021137f0d93cbaa48a24852eec74671a7dd8
Former-commit-id: 70822bac5dc2adc1b941b64b61e8375f14cdb747