Fix testPssh. *** DO NOT MERGE ***
The test is using content that does not have a clearkey uuid in
the pssh, but a Widevine uuid. So when it runs on platforms
without Widevine HIDL service, it will fail as it tries to
create a Widevine drm object to compare to the Widevine uuid in
the pssh.

This fix uses content that has a clearkey uuid in the pssh,
so the test no longer need to create a Widevine drm object,
but a clearkey drm object.

In addition,  this CL fixes the following issues:
1. Allow more time before checking if the lisenter has
   received EVENT_VENDOR_DEFINED event.
2. Fix b/62491118, fix incorrect path for checkCodecsForPath().

This fix is already in oc-mr1-dev, no need to merge.

Test: ANDROID_BUILD_TOP= ./android-cts/tools/cts-tradefed run cts -m
  CtsMediaTestCases --test android.media.cts.NativeClearKeySystemTest#testPssh

Test: ANDROID_BUILD_TOP= ./android-cts/tools/cts-tradefed run cts -m
  CtsMediaTestCases --test android.media.cts.NativeClearKeySystemTest

bug: 64161003
Change-Id: Id6d97498458772d5b91f669e082f7f3edb91b6de

Former-commit-id: 114c702c13a61438e84474b31e5f5fd2dc4f9a3a