Skip testAudioRoute if only one audio route
Currently, SelfManagedConnectionServiceTest.testAudioRoute will attempt
to change the audio route to the speaker and then to the earpiece. At
each step, it verifies that it receives a callback indicated that the
audio route has been changed. This makes the following assumptions:
  - Both ROUTE_SPEAKER and ROUTE_EARPIECE are available
  - ROUTE_SPEAKER is not currently being used for the call

On a phone, those should be good assumptions. However, on some devices
(e.g. watches), neither will be true. On watches, usually there will be
zero or one audio route available (SPEAKER). So, this test would always
fail. This CL modifies the test method to mimic ExtendedInCallServiceTest's
testSwitchAudioRoutes which will be skipped if either ROUTE_SPEAKER or
ROUTE_WIRED_OR_EARPIECE are missing.

Test: Executed test on angler and swordfish
Bug: 37683238
Change-Id: Id4c23e739f455057dbd87db5a4aa1a89c481e0c0

Former-commit-id: 41e417135d66fb2f6e5bec0e7802b6e3b305d458