[cts] Fix UsesLibraryTest
In order to ensure the collision check is executed we need to use the test
apk file when constructing the test class path for duplicates.

We cannot use the shared libraries apks because they are compiled with a
special marker which may skip the collision check (b/37777332).

Bug: 64988785
Bug: 37777332
Test: cts-tradefed run cts -m CtsAppSecurityHostTestCases --test
android.appsecurity.cts.UsesLibraryHostTest

Change-Id: If32ba7533258a73161aed997e6bfa8c5b6ba3ae7

Former-commit-id: 78695b9c1a038bb0df3e41409031d71c5b382cea