Fix flakyness of PreferenceRecycleTest#recycleSetViaCodeTest
We need to select new preferences added via code to make sure they're visible.
The selection may have refreshed the view or not, which makes the state of the
views inconsistent.

Simplify the test to remove intermediate check of the view status and only
verify that when views have been refreshed more than once, the recyclable one
should be recycled and the non-recyclable shouldn't.

Test: run cts on Huawei watch and Nexus 6P
Change-Id: I5fa999b2433e6b15104c09fe2f2002b2b3ee7992
FIXES: 65124578

Former-commit-id: 2c47acb964e07023ae7632ef111b4ade21fb187f