Allow For Separate Work Settings App In CTS-V
Update the wording of managed device tests that require a Settings app
to be opened, to allow the Work Settings app to be opened instead if
work and personal settings are split into different apps. Also change
cross profile intent test so some settings intents are allowed to either
be handled inside the work profile by a work settings app, or forwarded
to the primary profile.

Fixes: 64070330
Test: Run the relevant test cases in the CTS Verifier app

Former-commit-id: a54f629be711488467994916b3c733dd7d2ff134