Adapt PopupMenuTest#testGroupDividerEnabledAPI for smaller screens.
On smaller screens some of PopupMenu items are not visible and
do not have corresponding views.

Stop trying to access invisible items. Also anchor the menu to the
upper edge of the screen to make better use of the available room
(for this test and a similarly constructed
PopupMenuTest#testItemViewAttributes)

Bug: 68960209
Test: android.widget.cts.PopupMenuTest
Change-Id: Id977eecbd275e7ff7ba7de578e4ae4499895203d

Former-commit-id: e9720e4494b658cfccde2c57d4cb66d736b2076d