Updating CTS code to create user.
If maxuser is 2, CtsDevicePolicyManagerTestCases can not be passed.
Through the following command SecondaryUser will be created and one User will be added. (total: 2)
1. am instrument -w -r --user 0 -e timeout_msec 600000 -e class com.android.cts.comp.ManagementTest#testCreateSecondaryUser com.android.cts.comp/android.support.test.runner.AndroidJUnitRunner

In this state, it tries to create another user again, but since the current total user count is 2, it is equal to Max User Size, so user can not add User from UserManagerService and return error.
2. pm create-user --profileOf 0 --managed TestProfile_1509511814385

Test: manual

Change-Id: I05dac47e18c544ac9a867764e2a218c3b60f48ec
Signed-off-by: Yongjin Yeon <yongjin.yeon@lge.com>

Former-commit-id: 9f718381d71fb22e0f972b3bd89b65a56dac6a77