Rootless GPU Debug Tests
This adds a new set of tests that ensure debug Vulkan layers can be
enumerated and loaded from the base dir of debuggable apps.  It also
runs through several positive and negative scenarios to ensure the
new Settings work as expected.

The source builds multiple APKs: DEBUG and RELEASE for testing the
different combinations, and a LAYER APK just for transporting layers.

The layers themselves are null, only containing enough functionality
to be enumerated, chained, and loaded by the Vulkan loader.

Test: cts-tradefed run singleCommand cts -m CtsGpuToolsHostTestCases
Bug: 63708377

Change-Id: I5d1c7c005d4ff8f8629ad91b956d3769987f6170

Former-commit-id: fab47c3c3440911c6b9a69d101a5a5ecfc692381