Ensure that ActionBarActivity has focus before running MenuKey test.
There's a race where the activity has not gotten focus and sendKeyDownUpSync is
called. In this case; ActionBar's menu visibility listener does not
update even if the menu opens. We should check to make sure that the
activity has focus before checking for the menu to open.

Bug: 64489683
Test: Builds, ran android.app.cts.ActionBarTest on both Standlone Headset and
Sailfish and all tests pass.

Change-Id: Ie03b15557278b07affa0c2a9416ea1697119d9b2

Former-commit-id: 6cbcf0b0c0c23ae5d8d283e35846aeaf21b29cb6