Fixed test cases that rotated the device after the Save UI was shown.
These tests rotated the device back to portrait, which triggers a new session.

They used to pass because AutoFillServiceTestCase defines an @After
ignoreFurtherRequests() method that ignore them, but now somehow the timing has
changed so the new request is made before the @After rule is called

This CL properly fix the problem by explicitly waiting for that request.

Test: atest CtsAutoFillServiceTestCases:SimpleSaveActivityTest
Test: atest CtsAutoFillServiceTestCases

Fixes: 70348203

Change-Id: I11aca60bab80824f80aebf408ca243f6c2635dbf

Former-commit-id: db044c2f5356685569499ba99d854865c4b0fdb6