Fix display refresh rate test by removing default display assumption.
testRefreshRate checks the default display for refresh rate metrics.
While this works on normal devices this assumption breaks this test for
daydream standalone headsets because the vr-display is used to gather
fresh rate. This removes the assumption to use default display and uses
the appropriate display for both vr and non-vr devices.

Bug: 65742619
Test: CTS builds, test passes on both sailfish and daydream standalone
headset.

Change-Id: If6a87b6a7317cd3aafc681870e6c63c682fb6e3a

Former-commit-id: c11f5b5400372c065023e619a61c9b63e85d6671