Fix racy CTS tests involving phone state broadcast.
When registering the state listener in the CTS tests, TelephonyManager will
potentially send the starting state on registration, but in other cases
if TelephonyManager was already connected it won't sent the initial state.
Fixed this by comparing only the most recent state in the verify versus the
starting state.
Added some state logging just in case this isn't the actual problem.

Test: Ran unit tests; they pass.
Bug: 70530712
Change-Id: I27b255bc0d852d84faa5cd9e3b5a4735163e1683

Former-commit-id: 6fab333013314bd334b13c7e76d4f605c261655c