Make end-to-end IME tests more deterministic
In the previous implementation, when an end-to-end test in
CtsInputMethodTestCases is started it was possible that previously
running Activity can interact with MockIme until a TestActivity is
launched.

This CL introduces StateInitializeActivity so that each end-to-end
test cases in CtsInputMethodTestCases starts with showing a certain
behavior-defined Activity.  This should make CtsInputMethodTestCases
more deterministic and help us debug flaky failures.

Bug: 70730589
Test: CtsInputMethodTestCases
Change-Id: I9018d524cc4732418a6d77dc884fc416ab88f60a

Former-commit-id: 4a57cf66a84bf42515b02e1260651dd00eccdd40