Check if a dimming window cancels light nav bar
This is a preparation to allow IME windows to use
SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR.

One interesting behavior regarding light navigarion bar is that
showing a dimming window (a window that has FLAG_DIM_BEHIND) may
override SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR specified by a window
behind it.

To take this behavior into account, this CL checks if such a behavior
is supported or not by using a simple Activity.  If such a behavior is
actually supported, subsequent CTS test cases can rely on that
behavior and require that the IME windows behave in the same way.

Bug: 69002467
Test: CtsInputMethodTestCases
Change-Id: I2af6679a7f3d016d202f9e56a9c8442966ecd7cb

Former-commit-id: 374a412e2f9f9f1329af97b613cc16975f83d9d5