SystemApis are exact matched
SignatureTest now ensures that APIs annotated as @SystemAPI is not
altered from the documented API set. Specifically, SystemAPI annotation
is now changed to be retained at runtime and the test gathers all
classes, methods, fields, etc. that have the annotation. If some element
is found in the device which is missing in the documented API, then it
is reported as an error.

Bug: 67891551
Test: cts-tradefed run cts --module CtsSystemCurrentApiSignatureTestCases
shows no additional error
Test: cts/tests/signature/tests/run_unit_tests.sh

Change-Id: I4fbd03aadd609401b3e8bdd3b36165299e999261

Former-commit-id: 3e0fe7a955657794fbfb649535cc658c2b0f4d4e