DO NOT MERGE: Fix expected error for missing ZIP local header magic.
On N the error was "Failed to parse" and on O it is "Unknown failure".
On N-MR1 (7.1), since the change introduced the different is already
included there (hash: 4f26cb5a36487b3f5b497804faebe16241c15d54),
the error messages is "Unknown failure".

But either way, the error message is not an official part of APIs,
it is flagile for the test result to depend on it. Let's regard the
test as passing as long as the installation fails.

BUG=72130817
TEST=CtsAppSecurityHostTestCases

Change-Id: I08dd75d210449d4f87a8c2b0bf8cb0f50ccffd6f
(cherry picked from commit 48a7e352a13201266275a22ee60f2c27d6cd3137 [formerly 88c34e1595713ddb2be7506a7faf84af02c94550])

Former-commit-id: 66241a0b7d85a3ecd7c5885d71cd1f9d2f58f2fb