Skip CtsSyncAccountAccessOtherCertTestCases in test harness
The test harness presents a incorrectly behaving device. Hence tests
that rely on the correct behavior cannot be run.

In this case the test relies on the account manager to deny access to
accounts an adapter does not have access to. In the test harness all
access is alreays granted though.

Fixes: 72114924
Test: atest com.android.cts.content.testAccountAccess_otherCertAsAuthenticatorCanNotSeeAccount
      with ro.test_harness set and without

Change-Id: Ia5ea6e61d95bc238f952e8bdaff742a7bab6d504

Former-commit-id: 7802d25aaba73ba0501defc788cb9a30b2e92814