[AWARE] Remove "Respond to ANY" API - dummy tests deprecated
Aware API which permitted a Responder (Publisher) to accept data
path requests from ANY peer (as opposed to a specific peer with a
Peer ID or MAC address) is deprecated in the API update.

The CTS tests in any case cannot test the full exchange (which
requires 2 devices and is tested using CtsVerifier). They tested
the creation of a valid Network Specifier and a network request
which then times-out. However, the network request did not specify
a peer - a now deprecated API.

Update the tests:
- For OOB data-path specify a dummy MAC address
- Remove the in-band data-path tests since there is no public API
  to create a dummy PeerHandle.

Bug: 72175022
Test: CTS passes
Change-Id: I97b74d80d0f9dfdc43c4c743a18c960c05ab6a53

Former-commit-id: d7a77455b5262d906a0c6731b7208d8135269b14