Test font preloading with WebView.
Font preloading was interacting badly with WebView multiprocess and
causing a crash. To prevent this from regressing, specify a dummy
preloaded font array in CtsWebkitTestCases such that all of the WebView
tests in this APK will run the framework font preloading code (though
not actually load any fonts as the list is empty; the issue occurred
even with an empty list). Give the list resource a manually assigned
resource ID which is very unlikely to exist in the WebView
implementation APK so that the crash can't be avoided by chance.

Bug: 70968451
Test: CtsWebkitTestCases
Change-Id: Ib5463814e993641b6e32e551de4c0fd0627f5792

Former-commit-id: 896e9d7bc3b4ca3925812d865276fad59d7f32fe