CTS: When running in VR mode alert windows are ignored.
When running in VR mode a global App Op restriction is in place for
SYSTEM_ALERT_WINDOW. This has higher precedence than package level
access the test puts in place. The global App Op restriction causes
testModeSwitch to fail because the CtsTestPresentation is ingored,
being a SYSTEM_ALERT_WINDOW.

this patch bypasses testModeSwitch when running in VR mode.

Bug: 69856028
Bug: 68724913
Test: android.display.cts.DisplayTest#testModeSwitch passes on VR
      standalone VR device.

Change-Id: I1df9dce5416dae622b63a171717b19493467cce5

Former-commit-id: 3fab004e331fa6c5a6b255c2e175800676ee6c11