Temporarily disable WebView Tracing API CTS tests for landing
Temporarily disable existing TracingController CTS tests
in order to land the updated APIs (ag/3478990) and the
corresponding android webview glue and chromium patches.

The tests will be updated and re-enabled once the chromium
side rolls into android master and the test will actually
be able to run the new api end-to-end.

BUG: 63750258
Test: cts-tradefed run cts -m CtsWebkitTestCases --test android.webkit.cts.TracingControllerTest
Change-Id: I78f4ee09c5caf226c7275f5937407a3584450289

Former-commit-id: 02511db57c32a567644b3de0ccc33492e4441de1