Create CTS test for hidden API enforcement
This patch adds a CTS test which takes the generated blacklist of
private APIs and tries to discover them with reflection and JNI. If
the class member is discovered, it produces an error.

The notable omissions from this test are:
(a) need to also test dark greylist if target SDK version is >= P, and
(b) need to test accessing the APIs by statically linking against them.

The former depends on adding the corresponding logic to ART, the latter
on dynamically generating DEX bytecode.

Bug: 64382372
Test: cts/tests/signature/runSignatureTests.sh
Change-Id: I12cf06f0264fa195e1ba4666ca8903b5d8c0438d

Former-commit-id: 12f70c807383fccf5324d750ae5b971391a7ad32