WebView: fix WebSettingsTest for Safe Browsing by default
This test used to assert that Safe Browsing was off by default. Now that
Safe Browsing is enabled by default, this is no longer true.

We can't assert that getSafeBrowsingEnabled() returns true because it
also depends on user opt-in preference (and CTS always looks like a
user opt-out on GMS devices).

Bug: 67717640
Test: cts-tradefed run cts -m CtsWebkitTestCases -t android.webkit.cts.WebViewSettingsTest#testEnableSafeBrowsing
Change-Id: I07262e3a316c6d85c43e42c0708c16b68e82e3c8

Former-commit-id: 5965c8d76dab982df2ec49c2bfd019b616f1a402