DO NOT MERGE: Disable SignificantMotionTest on VR headset
The test currently has assumption that device must enter suspend mode
within 30 seconds of inactivity. On Daydream headset this interval is
increased to 5 minutes, but there is no API in API Level 26 to call
for the test to determine this timeout.

The plan is to add an API to future version of Android to query the
suspend timeout or to add a control to the test to specify the expected
suspend interval in the test UI. For now, skip this test on VR devices.

Bug: 70731002
Change-Id: Iaf362a32f8153acb8ea9f9751eb6712abf24df26

Former-commit-id: da4e391464dc3610123c3093aa1072c0da3bf295