Change CtsJdwpTestCases to run with new agent JDWP implementation.
This CTS test was still testing the old JDWP implementation. We should
be testing the new one. This test will look at the
dalvik.vm.jdwp-provider property and run the jdwp tests using
whichever jdwp implementation is being used. If the property is unset,
'default' or 'adbconnection' it will use the JVMTI agent jdwp
implementation. If it is 'internal' it will use the old internal ART
jdwp implementation.

Bug: 74028528
Bug: 62821960

Test: cts-tradefed run cts -m CtsJdwpTestCases
Test: adb setprop dalvik.vm.jdwp-provider internal &&
      cts-tradefed run cts -m CtsJdwpTestCases
Change-Id: I21aa701f30ff914be0350ff6db1692a849c88569

Former-commit-id: ced60ea170aa8ef7eac5e351f469a2c7a0e3d416