Add test to ensure there's only one app setting brightness curves.
If we have more than one app setting the brightness curve, whoever sets
it last wins, which is really undesirable since it might seem completely
random to the user. In addition, all the training being done to derive
the curve that isn't chosen is then just wasted work.

Also, move the tests to be full-fledged JUnit4 tests.

Bug: 74332874
Test: atest BrightnessTest
Change-Id: I98c84c2b13711e7227da0c3364b45b5b285905fa

Former-commit-id: 9962ce229eebc7c3b884580cd3b040611c283bc4