Allow device with preloaded accounts to run DO tests in CtsVerifier
Problem:
Some OEM have prealoaded accounts in their device, which makes
set-device-owner failed to run. To workaround this, we allow
set-device-owner to run if the target device admin is test-only.
However, we don't wanna to make the entire CtsVerifier app test-only,
as we may break some other tests.

Solution
Introduce a dummy device owner, and update the test instruction to ask
tester to set this new app to DO instead. This app monitors DO change,
and transfer the ownership to CtsVerifier automatically.

Test:
1. Add an account with DEVICE_OR_PROFILE_OWNER_ALLOWED ->
   try to run the set-device-owner command. Verify that CtsVerifier
   eventually becomes device owner. And tries to run a test that requires
   DO privilege.

FIXES: 64392045

Change-Id: I5cc5b417a5db9476a2b2d30315e5174f269a82d5

Former-commit-id: 8d2939250372bc25f44033a2166706c50e8fdacc