DO NOT MERGE Add delay for GamepadTestCase
According to uhid documentation, the proper usage for /dev/uhid is to
first call UHID_CREATE, then wait for UHID_START, and UHID_OPEN. After
the open is received, the device is ready to accept write events. Even
though our code follows this spec, sometimes the device just isnt ready
as soon as UHID_OPEN is received. That means that even though writing to
/dev/uhid succeeds, there are no events generated. This is also observed
by getevent, which shows that the device was added and removed, but no
events were produced.
The workaround is to add a small delay after registering the device
before writing the events.

This issue is only observed on pi-dev and cannot be reproduced on master
on the latest devices. Therefore, merge this only into pi-dev to unblock
the failing tests.
If the failure occurs again, it will be necessary to investigate this
further in the kernel.

Test: atest GamepadTestCase
Bug: 79169722
Change-Id: Ia3e71b172bc4a5b46accfce8be5d40dd698fb88a

Former-commit-id: 07ed0fc39fb6f3194a24b608a7197d7b45505eab