Fix concurrency issue in Telecom CTS test.
The verifyPhoneStateListenerCallbacksForCall method had the issue that it
would wait on the mCallbackSemaphore; this was problematic as that
semaphore triggers on the FIRST phone state update.  Its really the second
update that we want to wait on.

Added a "waitUntilConditionIsTrue" which will wait until we see the
state update which is desired and then continue verification.

Test: Ran CTS test; modified to force concurrency scenario.
Bug: 79372691
Change-Id: Ic5d986cfe3b023f81a7e59d8efd257c9f0d8aa17

Former-commit-id: df660fc2d32d988d1f9f332508462a6f123edf13