Wait for ScrollView on wear devices.
This isn't entirely ideal since it could be any ScrollView,
but currently there's no ID attached ScrollView of the dialog in
the copy in PackageInstaller or the core wear support lib (which
needs to be migrated to). Test can still fail but it's a lot less
frequent (9/10 local runs pass).

Bug: 80493264
Test: atest CtsAppSecurityHostTestCases:PermissionsHostTest#testDenialWithPrejudice23
Verified: on shiner. does not apply to phones

Change-Id: Ic85055ef093b187aadcb092f7812cbbefd4fd499

Former-commit-id: a78ae8747cd023d6d62294e24d92cad2a4ab1f41