Fix the icon pixels should be counted in.
It's not easy to use RGB color space to distinguish the icon part
and non-icon part. To use HSV color space to split the icon part
and non-icon part is easier than RGB color space. To add the the
condition that test the value of HSV color of pixel is between the
min and max HSV value.

Fixes: 112618204
Fixes: 112876178
Test: atest CtsSystemUiTestCases

Change-Id: I1d3c4b1e326dbca94c8e80632ed6ac6e8f70be55

Former-commit-id: 3e4e04e701245887e145217ae8c609e9f06a25af