AMTest: Do not launch Home activity if there is none
This CL does not launch a possible Home Activity the device declared
that it does not have one.

Certain devices, like ARC++ and Auto, do no have a functional Home Activity.
In the particular case of ARC++, it has a dummy Home Activity that is
being used because certain applications require one. But it is a
non-functional one: it doesn't have providers, icons, or anything. It is
just a dummy Activity.

If the dummy Home Activity is launched, then it will affect the test,
because even though it is a Dummy activity, from the point of view of
the ActivityManagerService, it is not a `isActivityTypeStandard()`
and affects the result of the test.

Bug: 116263424
test: atest android.app.cts.ActivityManagerTest
      passed both on ARC++ and phone devices
Change-Id: Ib231cdf78e150edad133a3349de1617bf6b705ef

Former-commit-id: 8d485dc2d39a125c2055069edcb5e2b77a6ed256