PixelCopyTest fails on ldpi devices
These are tests that rotate the screen with the Bitmap created on
the screen displayed.
However, in a terminal which customizes and increases the height
of the navigation bar, Drawing is not completed because the area
of ​​the screen to be drawn is insufficient.
After that, the test times out (fails).

We think that navigation bar is unnecessary in this test.
So we Added processing to hide the navigation bar of the
corresponding activity.

Bug: 123383181
Test: android.view.cts.PixelCopyTest PASS
Change-Id: I513ca7f8c6d1cedd0ab2fe35d08e8f1452f5ba25

Former-commit-id: ebc32c6078624f8f2d73956d4dd253307e4fec67