cts testFindNextFocusFromRect fix
Fix android.view.cts.FocusFinderTest#testFindNextFocusFromRect test. The
root cause is that the rect created is supposed to be on the border
between the top left and the top right buttons, but in the top left
button. Fix the test by creating a rect right in the middle of the top
left and the top right buttons.

Bug: 120920767
Fix: android.view.cts.FocusFinderTest#testFindNextFocusFromRect

Test: build succesfully.

Change-Id: I4d1dbd2ba984203ec681d57b3d32984c520b184e

Former-commit-id: ade3016e9a76702b00fef4abd3867db628a8a379